rename doc generator
