+++ templates/components/engines/typescript/agent/tools/document-generator.ts
